backend/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
praw==7.7.1
firebase-admin==6.2.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.25.2
feedparser==6.0.10
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
fastapi==0.104.1
uvicorn[standard]==0.24.0